feat(application): implementing episode search by series
This commit adds support for searching for series by partial name.

The user can now enter a partial name of a series, and the application will return a list of all series that match the partial name.

This feature is implemented by adding a new method to the `SeriesService` class called `getSeriesContainer`. This method takes a list of all series and a partial name as input, and returns an optional containing the first series that matches the partial name.

The `UserInterface` class is updated to use the new `getSeriesContainer` method to display a list of available series to the user.

This feature improves the user experience by making it easier for users to find the series they are looking for.